+++ android/src/main/java/com/getcapacitor/community/database/sqlite/SQLite/UtilsDownloadFromHTTP.java
@@ -1,13 +1,17 @@
 package com.getcapacitor.community.database.sqlite.SQLite;
 
+import java.net.MalformedURLException;
 import android.content.Context;
-import android.content.res.AssetManager;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.HttpURLConnection;
 import java.net.URL;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
 
 public class UtilsDownloadFromHTTP {
 
@@ -16,34 +20,46 @@ public class UtilsDownloadFromHTTP {
     private final UtilsFile _uFile = new UtilsFile();
 
     public void download(Context context, String fileUrl) throws Exception {
-        AssetManager assetManager = context.getAssets();
-        String fileName = fileUrl.substring(fileUrl.lastIndexOf('/') + 1);
-        String dbName = fileName;
-        Boolean isZip = false;
-        if (!fileName.contains("SQLite.db")) {
-            if (_uFile.getFileExtension((fileName)).equals("db")) {
-                dbName = fileName.substring(0, fileName.length() - 3) + "SQLite.db";
+         Boolean isZip = false;
+        String fileName = "";
+        try {
+            String[] fileDetails = getFileDetails(fileUrl);
+            fileName = fileDetails[0];
+            String extension = fileDetails[1];
+            if (!fileName.contains("SQLite.db")) {
+                switch(extension) {
+                    case "db":
+                        fileName = fileName.substring(0, fileName.length() - 3) + "SQLite.db";
+                        break;
+                    case "zip":
+                        isZip = true;
+                        break;
+                    default:
+                        throw new Exception("Unknown file type. Filename: " + fileName);
+                }
             }
-            if (_uFile.isLast(fileName, ".zip")) {
-                isZip = true;
-            }
+
+        } catch (Exception e) {
+            throw new Exception(e.getMessage());
         }
+
+
         File cacheDir = context.getCacheDir();
         String cachePath = cacheDir.getAbsolutePath();
-        String tmpFilePath = cachePath + File.separator + dbName;
+        String tmpFilePath = cachePath + File.separator + fileName;
         String databasePath = _uFile.getDatabaseDirectoryPath(context);
         File databaseDir = new File(databasePath);
         try {
             // delete file if exists in cache
             Boolean isExists = _uFile.isPathExists(tmpFilePath);
             if (isExists) {
-                _uFile.deleteFile(cachePath, dbName);
+                _uFile.deleteFile(cachePath, fileName);
             }
-            downloadFileToCache(fileUrl, cachePath);
+            downloadFileToCache(fileUrl, fileName, cachePath);
             if (isZip) {
                 _uFile.unzipCopyDatabase(cachePath, null, tmpFilePath, true);
                 // delete zip file from cache
-                _uFile.deleteFile(cachePath, dbName);
+                _uFile.deleteFile(cachePath, fileName);
             }
             // move files to database folder
             _uFile.moveAllDBs(cacheDir, databaseDir);
@@ -52,28 +68,41 @@ public class UtilsDownloadFromHTTP {
         }
     }
 
-    public static void downloadFileToCache(String fileURL, String cacheDir) throws Exception {
+    public static String[] getFileDetails(String url) throws Exception {
+        try {
+            URL javaUrl = new URL(url);
+            String path = javaUrl.getPath();
+            String decodedPath = URLDecoder.decode(path, StandardCharsets.UTF_8.toString()); // Decode URL-encoded path
+            String filename = decodedPath.substring(decodedPath.lastIndexOf('/') + 1); // Extract filename from decoded path
+            String extension = getFileExtension(filename);
+            if(extension == null) {
+                throw new Exception("extension db or zip not found");
+            }
+            return new String[]{filename, extension};
+        } catch (MalformedURLException e) {
+            e.printStackTrace();
+            throw new Exception(e.getMessage());
+        }
+    }
+    public static String getFileExtension(String filename) {
+        Pattern pattern = Pattern.compile("\\.([a-zA-Z0-9]+)(?:[\\?#]|$)");
+        Matcher matcher = pattern.matcher(filename);
+
+        if (matcher.find()) {
+            return matcher.group(1).toLowerCase(); // returns the matched extension in lowercase
+        }
+
+        return null; // no extension found
+    }
+    public static void downloadFileToCache(String fileURL, String fileName, String cacheDir) throws Exception {
         HttpURLConnection httpConn = null;
         try {
             URL url = new URL(fileURL);
             httpConn = (HttpURLConnection) url.openConnection();
+            httpConn.setRequestMethod("GET");
             int responseCode = httpConn.getResponseCode();
             if (responseCode == HttpURLConnection.HTTP_OK) {
-                String fileName = "";
-                String disposition = httpConn.getHeaderField("Content-Disposition");
-                //                String contentType = httpConn.getContentType();
                 int contentLength = httpConn.getContentLength();
-
-                if (disposition != null) {
-                    // extracts file name from header field
-                    int index = disposition.indexOf("filename=");
-                    if (index > 0) {
-                        fileName = disposition.substring(index + 10, disposition.length() - 1);
-                    }
-                } else {
-                    // extracts file name from URL
-                    fileName = fileURL.substring(fileURL.lastIndexOf("/") + 1);
-                }
                 String dbName = fileName;
                 if (!fileName.contains("SQLite.db")) {
                     if (fileName.substring(fileName.length() - 3).equals(".db")) {
@@ -81,26 +110,22 @@ public class UtilsDownloadFromHTTP {
                     }
                 }
 
-                //                System.out.println("Content-Type = " + contentType);
-                //                System.out.println("Content-Disposition = " + disposition);
-                //                System.out.println("Content-Length = " + contentLength);
-                //                System.out.println("fileName = " + fileName);
-                // opens input stream from the HTTP connection
-                InputStream inputStream = httpConn.getInputStream();
                 // create temporary file path
                 String tmpFilePath = cacheDir + File.separator + dbName;
-                // opens an output stream to save into file
-                FileOutputStream outputStream = new FileOutputStream(tmpFilePath);
-                int bytesRead = -1;
-                byte[] buffer = new byte[BUFFER_SIZE];
-                while ((bytesRead = inputStream.read(buffer)) != -1) {
-                    outputStream.write(buffer, 0, bytesRead);
+                // opens input stream from the HTTP connection
+                try (InputStream inputStream = httpConn.getInputStream();
+                     FileOutputStream outputStream = new FileOutputStream(tmpFilePath)) {
+
+                    byte[] buffer = new byte[1024];
+                    int bytesRead;
+                    while ((bytesRead = inputStream.read(buffer)) != -1) {
+                        outputStream.write(buffer, 0, bytesRead);
+                    }
+                    outputStream.close();
+                    inputStream.close();
+
+                    System.out.println("File " + fileName + " downloaded (" + contentLength + ")");
                 }
-
-                outputStream.close();
-                inputStream.close();
-
-                System.out.println("File " + fileName + " downloaded (" + contentLength + ")");
             } else {
                 String msg = "No file to download. Server replied HTTP code: " + responseCode;
                 throw new IOException(msg);