+++ ios/Plugin/CapacitorSQLite.swift
@@ -589,12 +589,9 @@ enum CapacitorSQLiteError: Error {
                 self.retHandler.rResult(call: call)
                 return
             case .failure(let error):
-
-                if error == .downloadFromHTTPFailed {
-                    let msg = "Download from HTTP failed"
-                    self.retHandler.rResult(call: call, message: msg)
-                    return
-                }
+                let msg = "Download from HTTP failed: \(error.localizedDescription)"
+                self.retHandler.rResult(call: call, message: msg)
+                return
 
             }
 
@@ -1013,6 +1010,7 @@ enum CapacitorSQLiteError: Error {
     // MARK: - deleteDatabase
 
     // swiftlint:disable cyclomatic_complexity
+    // swiftlint:disable function_body_length
     @objc func deleteDatabase(_ dbName: String, readonly: Bool) throws {
         guard isInit else {
             throw CapacitorSQLiteError.failed(message: initMessage)
@@ -1030,10 +1028,12 @@ enum CapacitorSQLiteError: Error {
         do {
             if !mDb.isDBOpen() {
                 // check the state of the DB
-                let state: State = UtilsSQLCipher.getDatabaseState(databaseLocation: databaseLocation, databaseName: "\(mDbName)SQLite.db", account: account)
-                if !isEncryption && 
-                        (state.rawValue == "ENCRYPTEDGLOBALSECRET" ||
-                         state.rawValue == "ENCRYPTEDSECRET") {
+                let state: State = UtilsSQLCipher.getDatabaseState(databaseLocation: databaseLocation,
+                                                                   databaseName: "\(mDbName)SQLite.db",
+                                                                   account: account)
+                if !isEncryption &&
+                    (state.rawValue == "ENCRYPTEDGLOBALSECRET" ||
+                    state.rawValue == "ENCRYPTEDSECRET") {
                     var msg = "Cannot delete an Encrypted database with "
                     msg += "No Encryption set in capacitor.config"
                     throw CapacitorSQLiteError.failed(message: msg)
@@ -1066,6 +1066,7 @@ enum CapacitorSQLiteError: Error {
             throw CapacitorSQLiteError.failed(message: msg)
         }
     }
+    // swiftlint:enable function_body_length
     // swiftlint:enable cyclomatic_complexity
 
     // MARK: - isJsonValid
@@ -1547,7 +1548,7 @@ enum CapacitorSQLiteError: Error {
             // get the database files
             let dbList: [String] = try UtilsFile.getFileList(path: aPath, ext: ".db")
             // filter for db including SQLite
-            let dbWithSQLite = dbList.filter { $0.contains("SQLite") }
+            let dbWithSQLite = dbList.filter({ $0.contains("SQLite") })
 
             return dbWithSQLite
 
@@ -1567,7 +1568,7 @@ enum CapacitorSQLiteError: Error {
             let dbList: [String] = try UtilsMigrate
                 .getMigratableList(folderPath: folderPath)
             // filter for db not including SQLite
-            let dbNoSQLite = dbList.filter { !$0.contains("SQLite") }
+            let dbNoSQLite = dbList.filter({ !$0.contains("SQLite") })
 
             return dbNoSQLite